[Enhancement][Olana] PR Comments
