+++ components/truncated-text.tsx
@@ -1,16 +1,17 @@
-import { Text } from "@ui-kitten/components";
+import { Text, useTheme } from "@ui-kitten/components";
 import React, { useState } from "react";
 import { TouchableOpacity } from "react-native";
 
-interface TruncatedTextInterface {
+interface PropsInterface {
     text: string;
     maxLength: number;
     onToggleExpanded?: (expanded: boolean) => void;
     color?: string;
 }
-export function TruncatedText(props: TruncatedTextInterface) {
+export function TruncatedText(props: PropsInterface) {
     const { text, maxLength, onToggleExpanded, color } = props;
     const [expanded, setExpanded] = useState(false);
+    const theme = useTheme();
 
     const toggleExpanded = () => {
         setExpanded(!expanded);
@@ -18,8 +19,7 @@ export function TruncatedText(props: TruncatedTextInterface) {
     };
 
     const displayedText = expanded ? text : text.slice(0, maxLength);
-
-    const textColor = color ? color : "#65676B";
+    const textColor = color ?? theme["color-gray-500"];
 
     return (
         <>
