+++ custom-theme.json
@@ -52,5 +52,6 @@
     "color-danger-600": "#DB282E",
     "color-danger-700": "#B71C2E",
     "color-danger-800": "#93112C",
-    "color-danger-900": "#7A0A2A"
+    "color-danger-900": "#7A0A2A",
+    "color-gray-500": "65676B"
 }
